Fixed bug where shader is used instead of program.
This was a stupid mistake on my part that took a while to debug. On
Linux and probably only a few OpenGL implementations the call to
glGetShader* on a linked shader program handle returns success (uh,
what!?) instead of crashing or returning an error code saying that
you're using the wrong function on the wrong handle somehow. Well, I
guess that's expected from OpenGL's error handling. What *really* annoys
me is that this behaviour is not standardized, i.e. if you pass wrong
stuff it should either return a failure or a success. There doesn't seem
to be a consensus on this accross implementations for some reason :-(.